refactor(cli): remove leftover references to deleted stream mirrors and child-stream machinery (#10915, #10906)
