+++ src/test-kernel/cli/ChildControls.vitest.ts
@@ -16,7 +16,7 @@ import {
   type StreamPhase,
   type StreamTabId,
 } from '@shared/schemas';
-import { buildChildRosters } from '@test/support/childStreamEntries';
+import { buildChildRosters } from '@test/support/childRosters';
 
 const root = 'root' as StreamTabId;
 const child = 'child' as StreamTabId;
