+++ src/test-kernel/cli/WorkflowScriptStreamTranscript.vitest.ts
@@ -584,7 +584,7 @@ describe('CLI workflow-script child-stream transcript', () => {
 
     const buildState = (repaintEpoch = 0) =>
       buildStaticTranscriptState({
-        childStreamEntries: new Map(),
+        childRosters: new Map(),
         maxRows: undefined,
         meta: SESSION_META,
         ownerKey: 'root',
@@ -598,7 +598,7 @@ describe('CLI workflow-script child-stream transcript', () => {
       current: ReturnType<typeof buildStaticTranscriptState>,
     ): ReturnType<typeof buildStaticTranscriptState> =>
       advanceStaticTranscriptState(current, {
-        childStreamEntries: new Map(),
+        childRosters: new Map(),
         maxRows: undefined,
         meta: SESSION_META,
         ownerKey: 'root',
@@ -1021,7 +1021,7 @@ describe('CLI workflow-script child-stream transcript', () => {
     ).toEqual([]);
 
     const staticItems = appendItems([], {
-      childStreamEntries: new Map([
+      childRosters: new Map([
         [
           PARENT_STREAM_ID,
           [
